refactor(tools): extract semantic_retriever as cross-cutting seam (arch PR A) (#317)
PR-A of the SymbolQuery pipeline series. `semantic_status` and
`semantic_results_for_query` previously lived inside `tools/symbols/
analyzer.rs` even though three impact-report families (boundary,
impact, refactor) plus the dispatch table consumed them through
`tools::symbols::*` re-exports. The symbol module was acting as an
incidental facade for a seam that does not belong to symbol queries.

Move them to a dedicated `tools/semantic_retriever` module — same
pattern as `tools/scip_health`. The seam now sits where every caller
imports from, and the symbol module no longer re-exports semantic
internals. Two adapters (symbol queries + impact reports) → a real
seam (Knuth: "one adapter = hypothetical seam, two = real seam").

No behaviour change — function bodies, cfg branches, and signatures
are identical. Internal `semantic_scores_for_query` in analyzer.rs
now routes through the new module. handlers.rs and dispatch/table.rs
import paths updated.

Sets up PR-B (SymbolQueryPipeline skeleton): the pipeline will
consume `semantic_retriever` like any other caller instead of owning
the semantic stage internally.

Build: cargo check --workspace --features semantic + --no-default-features
clean. cargo clippy --features semantic -D warnings clean. cargo fmt
clean. 7 existing symbols::tests pass.
