+++ crates/codelens-mcp/src/tools/mod.rs
@@ -16,6 +16,7 @@ pub mod reports;
 mod scip_health;
 pub(crate) mod semantic_edit;
 pub(crate) mod semantic_edit_args;
+pub(crate) mod semantic_retriever;
 pub mod session;
 mod suggestions;
 pub mod symbols;
