+++ crates/codelens-mcp/src/tools/reports/impact_reports/boundary.rs
@@ -2,7 +2,7 @@ use crate::AppState;
 use crate::tool_runtime::{ToolResult, required_string};
 use crate::tools::report_contract::make_handle_response;
 use crate::tools::report_utils::{stable_cache_key, strings_from_array};
-use crate::tools::symbols::{semantic_results_for_query, semantic_status};
+use crate::tools::semantic_retriever::{semantic_results_for_query, semantic_status};
 use codelens_engine::search::{SEMANTIC_COUPLING_THRESHOLD, SEMANTIC_NEW_RESULT_THRESHOLD};
 use codelens_engine::{find_circular_dependencies, get_change_coupling};
 use serde_json::{Value, json};
