+++ crates/codelens-mcp/src/tools/reports/impact_reports/refactor.rs
@@ -2,7 +2,7 @@ use crate::AppState;
 use crate::tool_runtime::ToolResult;
 use crate::tools::report_contract::make_handle_response;
 use crate::tools::report_utils::{stable_cache_key, strings_from_array};
-use crate::tools::symbols::{semantic_results_for_query, semantic_status};
+use crate::tools::semantic_retriever::{semantic_results_for_query, semantic_status};
 use serde_json::{Value, json};
 use std::collections::BTreeMap;
